- Fixed type problem in sctpsocket.cc. - Makefile.am update.
git-svn-id: https://svn.tdr.iem.uni-due.de:8443/svn/sctp/trunk/socketapi@1715 516812f1-d3fd-0310-9ea4-9aededebcb87

+++ socketapi/sctpsocket.cc
@@ -773,7 +773,7 @@ int SCTPSocket::internalReceive(SCTPNotificationQueue& queue,
          if(sda->sda_flags & SCTP_ARRIVE_UNORDERED) {
             flags |= MSG_UNORDERED;
          }
-         size_t receivedBytes = std::min((size_t) sda->sda_bytes_arrived, (size_t) bufferSize);
+         unsigned int receivedBytes = std::min((size_t) sda->sda_bytes_arrived, (size_t) bufferSize);
 #if (SCTPLIB_VERSION == SCTPLIB_1_0_0)
          unsigned int pathIndex;
          const int ok = sctp_receivefrom(assocID, streamID,
